Fix redefinition of ERROR. Fix expensive lookup for end of nr_t. Fix warnings on modern GCC about the use of %d to format size_t. Fix dquote compilation error on windows.

+++ include/env.h
@@ -207,7 +207,7 @@ struct track_t {
 typedef struct symrec_t symrec_t;
 struct symrec_t {
   token_t type;
-  size_t line, column;
+  uint32_t line, column;
   union {
     struct {
       char *identifier;
